refactor(core): move manifest parsing and RunMeta into leviath-core
Phase 4 of the engine-relocation refactor. Moves two pieces of pure logic
out of leviath-cli into leviath-core so a later phase can relocate the async
run engine into leviath-runtime without a dependency cycle. No new core
dependency (still serde/serde_json/thiserror/tracing/chrono/toml).

Move 1: manifest parsing
- New leviath_core::manifest module with the pure parse_manifest (string ->
  Blueprint) plus its parse_security_config helper and all pure unit tests.
- parse_manifest now returns leviath_core::Result (Error::Other) instead of
  anyhow, preserving the exact Display strings; anyhow was not pulled into core.
- find_manifest stays in cli (depends on the dirs crate); cli manifest.rs
  re-exports leviath_core::manifest::parse_manifest so internal call sites
  keep working.
- Deleted parse_manifest_public and its mod.rs re-export; all callers
  (test/list/pack/validate/dashboard/serve/integration) now call
  leviath_core::manifest::parse_manifest directly.

Move 2: run-state data types
- New leviath_core::run_meta module with the plain serde types RunMeta,
  RunStatus, ContextSnapshot, RegionSnapshot, RegionEntrySnapshot,
  StageRecord, StageRunStatus (plus new/touch/Display and a private now_secs).
- runstate.rs re-exports them so the ~192 cli references are unchanged; all
  filesystem IO free functions stay in cli.
